// src/generated/graphql.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** The `Long` scalar type represents 52-bit integers */
  Long: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};
// ...
export enum Enum_Shipping_City {
  Ca = 'CA',
  Ec = 'EC',
  Us = 'US'
}
// ...
export type MutationCreatePaymentIntentArgs = {
  amount: Scalars['Float'];
};
// ...
export type OrderInput = {
  cart?: InputMaybe<Scalars['ID']>;
  country?: InputMaybe<Scalars['String']>;
  delivered?: InputMaybe<Scalars['Boolean']>;
  email?: InputMaybe<Scalars['String']>;
  firstName?: InputMaybe<Scalars['String']>;
  lastName?: InputMaybe<Scalars['String']>;
  phoneNumber?: InputMaybe<Scalars['Long']>;
  publishedAt?: InputMaybe<Scalars['DateTime']>;
  shipping?: InputMaybe<Scalars['ID']>;
  state?: InputMaybe<Scalars['String']>;
  streetAdress?: InputMaybe<Scalars['String']>;
  zipCode?: InputMaybe<Scalars['String']>;
};

export type Pagination = {
  __typename?: 'Pagination';
  page: Scalars['Int'];
  pageCount: Scalars['Int'];
  pageSize: Scalars['Int'];
  total: Scalars['Int'];
};

export type PaginationArg = {
  limit?: InputMaybe<Scalars['Int']>;
  page?: InputMaybe<Scalars['Int']>;
  pageSize?: InputMaybe<Scalars['Int']>;
  start?: InputMaybe<Scalars['Int']>;
};

export type PaymentIntent = {
  __typename?: 'PaymentIntent';
  clientSecret?: Maybe<Scalars['String']>;
};
// ...
export type ResponseCollectionMeta = {
  __typename?: 'ResponseCollectionMeta';
  pagination: Pagination;
};

export type Shipping = {
  __typename?: 'Shipping';
  city?: Maybe<Enum_Shipping_City>;
  createdAt?: Maybe<Scalars['DateTime']>;
  price?: Maybe<Scalars['Float']>;
  publishedAt?: Maybe<Scalars['DateTime']>;
  updatedAt?: Maybe<Scalars['DateTime']>;
};

export type ShippingEntity = {
  __typename?: 'ShippingEntity';
  attributes?: Maybe<Shipping>;
  id?: Maybe<Scalars['ID']>;
};

export type ShippingEntityResponse = {
  __typename?: 'ShippingEntityResponse';
  data?: Maybe<ShippingEntity>;
};

export type ShippingEntityResponseCollection = {
  __typename?: 'ShippingEntityResponseCollection';
  data: Array<ShippingEntity>;
  meta: ResponseCollectionMeta;
};